test_main.py
"""
Test suite for the CreateQRCodeGUI class, which is responsible for creating a QR code GUI.

The tests cover various scenarios, including creating a QR code with a valid URL, a long URL, special characters, numeric input, and an empty input. The tests also check that the QR code is successfully created and saved, and that the appropriate label text is displayed.
"""
tests/test_main.py
class TestCreateQRCodeGUI(unittest.TestCase):

    def setUp(self):
        self.root = tk.Tk()
        self.s = tk.StringVar()
        self.entry = tk.Entry(self.root, textvariable=self.s)
        self.label = tk.Label(self.root)

    def tearDown(self):
        self.root.destroy()

    @patch('PIL.ImageTk.PhotoImage')
    def test_create_qrcode_with_image(self, mock_photo_image):
        mock_photo_image.return_value = MagicMock()
        self.s.set('https://example.com')
        create_qrcode()
        self.assertTrue(mock_photo_image.called)
        self.assertEqual(self.label.cget('text'), 'QR Code is created and saved successfully')

    def test_create_qrcode_with_long_input(self):
        long_input = 'https://example.com/this/is/a/very/long/url/that/exceeds/the/maximum/length/allowed/for/qr/codes'
        self.s.set(long_input)
        create_qrcode()
        self.assertEqual(self.label.cget('text'), 'QR Code is created and saved successfully')

    def test_create_qrcode_with_special_characters(self):
        input_with_special_chars = 'https://example.com/special_chars?#@!$%^&*()_+='
        self.s.set(input_with_special_chars)
        create_qrcode()
        self.assertEqual(self.label.cget('text'), 'QR Code is created and saved successfully')

    def test_create_qrcode_with_numeric_input(self):
        numeric_input = '12345678901234567890'
        self.s.set(numeric_input)
        create_qrcode()
        self.assertEqual(self.label.cget('text'), 'QR Code is created and saved successfully')

    def test_create_qrcode_with_empty_input(self):
        self.s.set('')
        create_qrcode()
        self.assertEqual(self.label.cget('text'), 'QR Code is created and saved successfully')
